+++ src/lsp-client/LSPServersManager.h
@@ -20,7 +20,7 @@ public:
 	virtual ~LSPServerConfigInterface() = default;
 	virtual const bool IsFileTypeSupported (const BString& fileType) const = 0;
 	const char* const* Argv() const { return fArgv.data(); }
-				int32  Argc() const { return fArgv.size(); }
+				int32  Argc() const { return int32(fArgv.size()); }
 
 protected:
 	std::vector<const char*> fArgv;
